+++ crates/pallets/starknet/src/lib.rs
@@ -169,6 +169,8 @@ pub mod pallet {
         type ChainId: Get<Felt252Wrapper>;
         #[pallet::constant]
         type MaxRecursionDepth: Get<u32>;
+        #[pallet::constant]
+        type ProgramHash: Get<Felt252Wrapper>;
     }
 
     /// The Starknet pallet hooks.
@@ -1211,6 +1213,11 @@ impl<T: Config> Pallet<T> {
     pub fn chain_id() -> Felt252Wrapper {
         T::ChainId::get()
     }
+
+    pub fn program_hash() -> Felt252Wrapper {
+        T::ProgramHash::get()
+    }
+
     pub fn is_transaction_fee_disabled() -> bool {
         T::DisableTransactionFee::get()
     }