+++ starknet-e2e-test/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod ethereum_sandbox;
+pub mod starknet_contract;